+++ plugins/nf-co2footprint/src/main/nextflow/co2footprint/DataContainers/CIDataMatrix.groovy
@@ -1,7 +1,7 @@
 package nextflow.co2footprint.DataContainers
 
 import groovy.util.logging.Slf4j
-import nextflow.co2footprint.utils.HelperFunctions
+
 import nextflow.co2footprint.Logging.Markers
 
 import java.nio.file.Path