+++ plugins/nf-co2footprint/src/main/nextflow/co2footprint/FileCreation/SummaryFileCreator.groovy
@@ -5,7 +5,7 @@ import groovyx.gpars.agent.Agent
 import nextflow.co2footprint.CO2FootprintComputer
 import nextflow.co2footprint.Records.CO2EquivalencesRecord
 import nextflow.co2footprint.CO2FootprintConfig
-import nextflow.co2footprint.utils.Converter
+import nextflow.co2footprint.Metrics.Converter
 import nextflow.trace.TraceHelper
 
 import java.nio.file.Path
@@ -60,9 +60,9 @@ class SummaryFileCreator extends BaseFileCreator {
 
         String outText = """\
         Total CO2e footprint measures of this workflow run (including cached tasks):
-          CO2e emissions: ${Converter.toReadableUnits(totalStats['co2e'],'m', 'g')}
-          Energy consumption: ${Converter.toReadableUnits(totalStats['energy'],'m', 'Wh')}
-          CO2e emissions (market): ${totalStats['co2eMarket'] ? Converter.toReadableUnits(totalStats['co2eMarket'], 'm', 'g') : "-"}
+          CO2e emissions: ${Converter.toReadableUnits(totalStats['co2e'],'', 'g')}
+          Energy consumption: ${Converter.toReadableUnits(totalStats['energy'],'k', 'Wh')}
+          CO2e emissions (market): ${totalStats['co2eMarket'] ? Converter.toReadableUnits(totalStats['co2eMarket'], '', 'g') : "-"}
 
         """.stripIndent()
         List<String> readableEquivalences = equivalences.getReadableEquivalences()