+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Added `Markers.silentUnique` to deduplicate messages without prepending a prefix
 - Content of total CO2e metrics now in tabs instead of <details> block
 - HTML & CSS clarity improvements through comments and block separation
+- Removed some unnecessary conversions in the Computer, JavaScript and Tests
+- Removed `utils` and `HelperFunctions` in favor of `Metrics` class
 
 ## Features:
 - Base config class with more control over parameters